FEAT: ADD playGame function for the game logic
Main function for the game logic. Uses a global ROUNDS constant
for the number of rounds. Loop ROUNDS number of times
to call the respective functions for getting the human and
computer choices, and executes a round. After the completion
of the loop, declares a winner is there is one, or tie instead